The Election Data is being read in
WHAT:
Everything within the slashed lines.
The program now reads in the election data for the various countries. It discerns between the different parties names, and saves the numbers gathered from the data as separate, aptly named ints.

IMPORTANCE:
In order to accurately make the map purple, or any color at all, we need the election stats. Now that the program reads in the stats, we can use them to calculate a number.

WHY I DID IT THE WAY I DID IT:
The file reads in a a line of code, with no spaces to separate. The only thing that consistently separates the numbers is commas, so I used every comma as a space. Whenever the code comes across a comma, one of the if/else if/else conditionals are initiated, and that is categorized as either democratic, republican, or independent, based on their position in the String. If that county isn't recorded, then it is skipped.

NOTE: The program currently does not do anything with this data.
NOTE: Not all of the states (i.e. DC) has election data for every county.
NOTE: It is assumed all of the data sets are similar, and the election results are read in: republican --> democratic --> independent.
SLIGHTLY LESS IMPORTANT NOTE: It also doesn't read anything on the first line of the project.