Buttons
WHAT:
Now there are buttons for states vs. counties, and the election years. When the user clicks on the buttons, then the program can determine exactly which button the user has pressed. If the user presses on either the county/states buttons, then the current map is erased, and the new map that they desired is drawn. The election buttons have the potential to do something, but currently don't do anything. The code is well commented.

IMPORTANCE:
This is a massive milestone. Now the user has some input, and we can easily implement the election years in the same way. Once we convert the election data into a color palette, we can easily implement those into the buttons. The color changes have not been done yet.

WHY I DID IT THE WAY I DID IT:
It is the simplest way, and allows for improvement upon in the future.

NOTE: The buttons still don't actually change the colors of the map (colors aren't programmed in), but they have the potential.
NOTE: The map could potentially shrink a few inches, to make the map and the buttons fit better together. Not essential (or necessary), though.