My Added Functionality Improved.
WHAT:
The program can interpret the first line of code in the election data. A box is drawn that will contain my graphs, as well as the outline of the graphs and the titles of the different parties in their respective colors. 

IMPORTANCE:
This is the basis of my individual functionality, but ultimately has no bearing on the rest of the project.

WHY I DID IT THE WAY I DID IT:
The file reads in a a line of code, with no spaces to separate. The only thing that consistently separates the numbers is commas, so I used every comma as a space. Whenever the code comes across a comma, one of the if/else if/else conditionals are initiated, and that is categorized as either democratic, republican, or independent, based on their position in the String.

NOTE: Nothing is commented