My Added Functionality Improved.
WHAT:
The program compares the locaction of the mouse to the points in the polygon. No apparant correlations. About to try something else.

IMPORTANCE:
This is more of my individual functionality, but ultimately has no bearing on the rest of the project.

WHY I DID IT THE WAY I DID IT:
First thing I thought of

NOTE: Nothing is commented