My Added Functionality Improved.
WHAT:
The program prints which state it's on, and can 'discern' between the states. Unfortunately, it believes that every point is either Idaho, California, or (occassionally) Alabama.

IMPORTANCE:
This is more of my individual functionality, but ultimately has no bearing on the rest of the project.

WHY I DID IT THE WAY I DID IT:
First thing I thought of

NOTE: Nothing is commented