My Added Functionality Improved.
WHAT:
MUCH more precision now, when telling where the mouse it (some of the states are actually accurate) 

IMPORTANCE:
This is more of my individual functionality, but ultimately has no bearing on the rest of the project.

WHY I DID IT THE WAY I DID IT:
First thing I thought of

NOTE: Nothing is commented
NOTE: the program is still confused, & thinks everything is Idaho/California, but getting better.