My Added Functionality Improved.
WHAT:
Commented out a lot of extraneous, pre-commented out debugging code. The graph is more precise (can tell which general state it's in), but doesn't adhere closely to the borders of the states.

IMPORTANCE:
This is more of my individual functionality, but ultimately has no bearing on the rest of the project.

WHY I DID IT THE WAY I DID IT:
First thing I thought of

NOTE: Nothing is commented
NOTE: the program is still confused, & thinks everything is Idaho/California, but getting better.