My Added Functionality Improved.
WHAT:
SO close to getting the recognition perfected. It's a ton better than before.

IMPORTANCE:
This is more of my individual functionality, but ultimately has no bearing on the rest of the project.

WHY I DID IT THE WAY I DID IT:
First thing I thought of

NOTE: Nothing is commented
NOTE: the program is still confused, but getting better.