My Added Functionality + Modularization + Comments.
WHAT:
The entirety of the new code is commented. The program is now modularized. My added functionality can usually locate the user's position within the map. The program can correlate the election data to a rough plotted graph, which changes as the user's mouse moves.

IMPORTANCE:
This is mostly pertaining to my individual functionality, but ultimately has no bearing on the rest of the project. The only use to the rest of the project is the modularizing the program.

WHY I DID IT THE WAY I DID IT:
The first, simplest thing I could think of

NOTE: the program is still confused when it comes to California/Nevada, Virginia/West Virginia, and Colorado/Alabama.