My Added Functionality
WHAT:
The problems with California/Nevada and Virginia/West Virginia are fixed.

IMPORTANCE:
This is mostly pertaining to my individual functionality, but ultimately has no bearing on the rest of the project. The only use to the rest of the project is the modularizing the program.

WHY I DID IT THE WAY I DID IT:
The first, simplest thing I could think of

NOTE: the program is still confused when it comes to Colorado/Alabama.