My Added Functionality
WHAT:
The graph plots the points relative to the data input

IMPORTANCE:
This is mostly pertaining to my individual functionality. The only use to the rest of the project is the modularizing the program.

WHY I DID IT THE WAY I DID IT:
The first, simplest thing I could think of