My Added Functionality
WHAT:
Everything is perfect. The graph plots the points relative to the data input, and the Independent works now (apparently, it didn't before). The colors work, the graph is scaled, and all that's left is commenting.

IMPORTANCE:
This is mostly pertaining to my individual functionality. The only use to the rest of the project is the modularizing the program.

WHY I DID IT THE WAY I DID IT:
The first, simplest thing I could think of