My Added Functionality
WHAT:
Some of the unused variables/methods have been removed. Three of the four classes have been completely commented.

IMPORTANCE:
This is mostly pertaining to my individual functionality. The only use to the rest of the project is the modularizing the program.

WHY I DID IT THE WAY I DID IT:
The first, simplest thing I could think of