My Added Functionality
WHAT:
Some more of the unused variables/methods have been removed. All of the classes have been completely commented.

IMPORTANCE:
This is mostly pertaining to my individual functionality. The only use to the rest of the project is the modularizing the program.

WHY I DID IT THE WAY I DID IT:
The first, simplest thing I could think of