Added Functionality + Color Fix
WHAT:
My added functionality is complete: the graph can read in the mouse from anywhere on the states. The buttons indicate which election year the user is looking at.  The problems with the colors class have been corrected. The index and titles have been added to my graph. Some more of the unused variables/methods have been removed. All of the classes have been completely commented.

IMPORTANCE:
This is the final version of the project. This encompasses everything that we were required to do in the project.

WHY I DID IT THE WAY I DID IT:
It got us to the end goal.